+++ src/Console/Commands/Game/RoundOpenCommand.php
@@ -5,6 +5,7 @@ namespace OpenDominion\Console\Commands\Game;
 use Carbon\Carbon;
 use Illuminate\Console\Command;
 use OpenDominion\Console\Commands\CommandInterface;
+use OpenDominion\Factories\RealmFactory;
 use OpenDominion\Factories\RoundFactory;
 use OpenDominion\Models\RoundLeague;
 use RuntimeException;
@@ -20,10 +21,13 @@ class RoundOpenCommand extends Command implements CommandInterface
                              {--realmSize=10 : Maximum number of dominions in one realm}
                              {--packSize=4 : Maximum number of players in a pack}
                              {--playersPerRace=2 : Maximum number of players using the same race, 0 = unlimited}
-                             {--mixedAlignment : Allows for mixed alignments}';
+                             {--mixedAlignment=true : Allows for mixed alignments}';
 
     /** @var string The console command description. */
     protected $description = 'Creates a new round which starts in 5 days';
+
+    /** @var RealmFactory */
+    protected $realmFactory;
 
     /** @var RoundFactory */
     protected $roundFactory;
@@ -32,12 +36,16 @@ class RoundOpenCommand extends Command implements CommandInterface
      * RoundOpenCommand constructor.
      *
      * @param RoundFactory $roundFactory
+     * @param RealmFactory $realmFactory
      */
-    public function __construct(RoundFactory $roundFactory)
-    {
+    public function __construct(
+        RoundFactory $roundFactory,
+        RealmFactory $realmFactory
+    ) {
         parent::__construct();
 
         $this->roundFactory = $roundFactory;
+        $this->realmFactory = $realmFactory;
     }
 
     /**
@@ -74,7 +82,7 @@ class RoundOpenCommand extends Command implements CommandInterface
             throw new RuntimeException('Option --realmSize must be greater than or equal to option --packSize.');
         }
 
-        if($playersPerRace < 0) {
+        if ($playersPerRace < 0) {
             throw new RuntimeException('Option --playersPerRace must be greater than or equal to 0.');
         }
 
@@ -96,12 +104,38 @@ class RoundOpenCommand extends Command implements CommandInterface
         }
 
         $startDate = new Carbon($startDate);
+
+        /** @var RoundLeague $roundLeague */
         $roundLeague = RoundLeague::where('key', $league)->firstOrFail();
 
         $this->info("Starting a new round in {$roundLeague->key} league");
 
-        $round = $this->roundFactory->create($roundLeague, $startDate, $realmSize, $packSize, $playersPerRace, $mixedAlignments);
+        $round = $this->roundFactory->create(
+            $roundLeague,
+            $startDate,
+            $realmSize,
+            $packSize,
+            $playersPerRace,
+            $mixedAlignments
+        );
 
         $this->info("Round {$round->number} created in {$roundLeague->key} league, starting at {$round->start_date}. With a realm size of {$round->realm_size} and a pack size of {$round->pack_size}");
+
+        if ($round->mixed_alignment) {
+            // Prepopulate round with 20 mixed realms
+            for ($i = 1; $i <= 20; $i++) {
+                $realm = $this->realmFactory->create($round);
+                $this->info("Realm {$realm->name} (#{$realm->number}) created in Round {$round->number} with an alignment of {$realm->alignment}");
+            }
+        } else {
+            // Prepopulate round with 5 good and 5 evil realms
+            for ($i = 1; $i <= 5; $i++) {
+                $realm = $this->realmFactory->create($round, 'good');
+                $this->info("Realm {$realm->name} (#{$realm->number}) created in Round {$round->number} with an alignment of {$realm->alignment}");
+
+                $realm = $this->realmFactory->create($round, 'evil');
+                $this->info("Realm {$realm->name} (#{$realm->number}) created in Round {$round->number} with an alignment of {$realm->alignment}");
+            }
+        }
     }
 }