+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes relevant to players in this project will be documented in th
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 
 ## [Unreleased]
+### Changed
+- A bunch of empty realms now get created for each new round, to prevent people landing together when not packing
+
+### Fixed
+- Units in training now count towards max military population
 
 ## [0.7.0-2] - 2019-08-09
 ### Added